fix(render): store bone indices as float so multi-bone bbmodels render
The custom BoneIndices vertex attribute was registered as
VertexFormatElement.Type.INT (4x int, shader input `in ivec4 BoneIndices`).
On the GPU skinning path this attribute was read out-of-range, so the
ksg_readBoneTransform texelFetch into the bone TBO returned zero, the skin
matrix collapsed to zero and every skinned vertex fell onto the origin,
hiding the whole model. Only single-bone models (bone index always 0) kept
rendering, which is why models with named groups disappeared. CPU skinning
does not read this attribute and was unaffected.

Fix:
- RenderState: register BONE_INDICES as Type.FLOAT and keep the 3-byte
  Normal last so the remaining attributes stay 4-byte aligned.
- FlatModelData: write bone indices with putFloat.
- ksg_skinning.transform.glsl / ksglib_main.vsh / ksglib_global_batch.vsh:
  declare BoneIndices as vec4 and round with int(BoneIndices[i] + 0.5)
  in BDEF/QDEF/SDEF.
- TransformFeedbackProgram: use a float pointer for the Iris
  transform-feedback source BoneIndices attribute.

Fixes #48

+++ modules/modelling/src/main/java/lib/kasuga/rendering/models/mc/backend/FlatModelData.java
@@ -964,10 +964,10 @@ public class FlatModelData implements AutoCloseable {
         int weightPos = bufPos + bindingWeightOffset;
         for (int i = 0; i < 4; i++) {
             if (i >= weightCount) {
-                buffer.putInt(indexPos + 4 * i, 0);
+                buffer.putFloat(indexPos + 4 * i, 0f);
                 buffer.putFloat(weightPos + 4 * i, 0f);
             } else {
-                buffer.putInt(indexPos + 4 * i, vertexBones[weightOffset + i]);
+                buffer.putFloat(indexPos + 4 * i, vertexBones[weightOffset + i]);
                 buffer.putFloat(weightPos + 4 * i, weights[weightOffset + i]);
             }
         }
